+++ NEWS.md
@@ -1,3 +1,14 @@
+# editData 0.1.1
+==================
+(2017-Oct-3)
+
+* Rewrite the R codes for 'editData' function as shiny modules
+
+* New shiny module functions "editableDTUI" and "editableDT" added. These module are intended for reuse across applications. 
+
+* New function "textInput3","seletInput3","label3","numericInput3","checkboxInput3",
+"radioButtons3" and "dateInput3" added for side-by-side input
+ 
 # editData 0.1.0
 ====================
 (2017-Sep-22)